Refactors the routing logic
This commit refactors the routing logic to switch between the shop and the cart. React-router was intended to use at the begining but I was not able because of the class components use (in which I am not used to). Opt for a simple way to switch this with a change of state and a button